+++ desktop/cmp/input/JsonInput.ts
@@ -11,7 +11,6 @@ import * as codemirror from 'codemirror';
 import 'codemirror/mode/javascript/javascript';
 import {codeInput, CodeInputProps} from './CodeInput';
 import {jsonlint} from './impl/jsonlint';
-import {isPlainObject} from 'lodash';
 
 export type JsonInputProps = CodeInputProps;
 
@@ -22,20 +21,11 @@ export const [JsonInput, jsonInput] = hoistCmp.withFactory<JsonInputProps>({
     displayName: 'JsonInput',
     className: 'xh-json-input',
     render(props, ref) {
-        let {value, ...codeInputProps} = props;
-
-        if (isPlainObject(value)) {
-            value = JSON.stringify(value, null, 2);
-        } else if (value === null) {
-            value = '';
-        }
-
         return codeInput({
             linter,
             formatter: fmtJson,
             mode: 'application/json',
-            ...codeInputProps,
-            value,
+            ...props,
             ref
         });
     }